update Autowired to Resource

+++ tcgl-modules/tcgl-system/src/main/java/com/tcgl/system/mapper/SysDeptMapper.java
@@ -27,7 +27,7 @@ public interface SysDeptMapper extends BaseMapper<SysDept> {
      * @param deptCheckStrictly 部门树选择项是否关联显示
      * @return 选中部门列表
      */
-    List<Integer> selectDeptListByRoleId(@Param("roleId") Long roleId, @Param("deptCheckStrictly") boolean deptCheckStrictly);
+    List<Long> selectDeptListByRoleId(@Param("roleId") Long roleId, @Param("deptCheckStrictly") boolean deptCheckStrictly);
 
     /**
      * 根据部门ID查询信息
